move mock server to testing
